+++ tools/project_template/ModuleTemplate/trace.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include <common/Telemetry/TraceBase.h>
-
-class Trace : public telemetry::TraceBase
+class Trace
 {
 public:
+    static void RegisterProvider();
+    static void UnregisterProvider();
     static void MyEvent();
 };